Migrating Error type to own module
